5.28 勝利代碼
yeah，但是只做了第二问，第一问差点GG，辛亏老师贼拉善良

+++ stm32/Hardware/encoder.h
@@ -9,4 +9,4 @@ void TIMER_config2(void);
 void Encoder_Init(void);
 void TIMER_config1(void);
 void Encoder_Display_Content(void);
-void Encoder_Delay(short marks);
+void Encoder_Delay(int marks);
